Optimize maps, smaller mapsize

+++ Frutty/src/frutty/Main.java
@@ -55,7 +55,7 @@ public class Main {
 	public static String loadMapSize(String fileName) {
 		try(ObjectInputStream input = new ObjectInputStream(new FileInputStream("./maps/" + fileName + ".deg"))){
 			input.readUTF();
-			return input.readInt() + "x" + input.readInt();
+			return input.readShort() + "x" + input.readShort();
 		} catch (IOException e) {}
 		return null;
 	}
